# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: QA Automation Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Projects
...
  /scheduled-audits:
    get:
      operationId: listScheduledAudits
      tags: [scheduled-audits]
      summary: List scheduled audits
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of scheduled audits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduledAudit"
    post:
      operationId: createScheduledAudit
      tags: [scheduled-audits]
      summary: Create a scheduled audit
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduledAuditInput"
      responses:
        "201":
          description: Created scheduled audit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduledAudit"
  /scheduled-audits/{id}:
    get:
      operationId: getScheduledAudit
      tags: [scheduled-audits]
      summary: Get scheduled audit by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scheduled audit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduledAudit"
    patch:
      operationId: updateScheduledAudit
      tags: [scheduled-audits]
      summary: Update a scheduled audit
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduledAuditUpdate"
      responses:
        "200":
          description: Updated scheduled audit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduledAudit"
    delete:
      operationId: deleteScheduledAudit
      tags: [scheduled-audits]
      summary: Delete a scheduled audit
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Notifications
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, tester]
        createdAt:
          type: string
          format: date-time
    UserInput:
      type: object
      required: [email, name, role, password]
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [admin, tester]
        password:
          type: string
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [admin, tester]
        password:
          type: string

    # Projects
...
    ScheduledAudit:
      type: object
      required: [id, name, projectId, frequency, hour, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        projectId:
          type: integer
        projectName:
          type: ["string", "null"]
        createdById:
          type: ["integer", "null"]
        frequency:
          type: string
          enum: [daily, weekly, monthly]
        hour:
          type: integer
        status:
          type: string
          enum: [active, paused, disabled]
        nextRunAt:
          type: ["string", "null"]
          format: date-time
        lastRunAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    ScheduledAuditInput:
      type: object
      required: [name, projectId, frequency, hour]
      properties:
        name:
          type: string
        projectId:
          type: integer
        frequency:
          type: string
          enum: [daily, weekly, monthly]
        hour:
          type: integer
        status:
          type: string
          enum: [active, paused, disabled]
    ScheduledAuditUpdate:
      type: object
      properties:
        name:
          type: string
        frequency:
          type: string
          enum: [daily, weekly, monthly]
        hour:
          type: integer
        status:
          type: string
          enum: [active, paused, disabled]

    # Notifications
...